pyproject.toml
[tool.black]
line-length = 79

include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

experimental-string-processing = true

[tool.isort]
line_length = 79

multi_line_output = 3

force_single_line = true

sections = ['FUTURE', 'STDLIB', 'THIRDPARTY', 'FIRSTPARTY', 'LOCALFOLDER']

profile = "black"


[tool.pylint]
disable = [
    "redefined-outer-name",
]


[tool.poetry]
name = "moneyflowbot"
version = "0.1.0"
description = "telegram API for MoneFlow"
authors = [ "Nick-voz <nik.vozisow@gmail.com>" ]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^21.6"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
